Keep observation data as DataFrame, not sparse
Instead of converting observation data to coo arrays, which are hard to work with, we
keep the data on DataFrame form. This makes it much easier to work with, and more general.
The user may still convert to coo_array or something else if needed.

This also renders train/test splitting and random subsampling of all-normals, even for only subsets, much easier.